🚚 rename: 모킹 역할을 수행하는 의도를 드러낼 수 있도록 접미사 붙임 #6
